Move identical functions up to base WebClient instance.
